refactor(main): extract title-bar popup into popups/titlePopup
Amp-Thread-ID: https://ampcode.com/threads/T-019e20b6-b1c6-76cf-9baf-54176cc188d1
